weather_ml/requirements.txt
# Weather ML Pipeline — Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
requests>=2.31.0
joblib>=1.3.0
